+++ packages/virtual/core/src/index.ts
@@ -3,6 +3,8 @@
  * {@inheritDoc VirtualScroller}
  */
 
+import "@af-utils/scrollend-polyfill";
+
 export { default as VirtualScroller } from "models/VirtualScroller";
 export { VirtualScrollerEvent, _ALL_EVENTS } from "constants/";
 export * from "./types";