+++ packages/virtual/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/virtual-core
 
+## 0.0.4
+
+### Patch Changes
+
+-   upgraded scrollToIndex, improved docs, added scrollend polyfill
+-   Updated dependencies
+    -   @af-utils/scrollend-polyfill@0.0.2
+
 ## 0.0.3
 
 ### Patch Changes